pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "greenscreen"
version = "0.2.0"
description = "Green screen video to alpha PNG sequences, GIF, and video (Lens Studio friendly)"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "antoinebou13" }]
dependencies = [
    "typer>=0.12",
    "rich>=13.7",
    "tqdm>=4.66",
    "pillow>=10.0",
    "numpy>=1.26",
    "opencv-python>=4.9",
    "natsort>=8.4",
    "imageio>=2.33",
    "imageio-ffmpeg>=0.4.9",
    "numpngw>=0.1.2",
    "yt-dlp>=2024.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.8",
]

[project.scripts]
greenscreen = "greenscreen.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --strict-markers"
